rename ConstantReplacerRector to RenameConstantRector
